# docs/openapi.yaml
openapi: 3.0.3
info:
  title: "Онлайн-кинотеатр API"
  description: |
    Полноценный API для стримингового сервиса с поддержкой фильмов, сериалов, рейтингов и пользовательских закладок.
    
    ## Основные возможности:
    - 📺 Просмотр фильмов и сериалов
    - ⭐ Система рейтингов и отзывов
    - 📋 Личные закладки пользователей
    - 👥 Управление пользователями и подписками
    - 📊 Аналитика и статистика
    - 🔧 Управление справочниками (админ)
    
    ## Аутентификация:
    Все защищенные эндпоинты требуют JWT токен, полученный через `/auth/login`.
    
    ## Технологии:
    - FastAPI 0.104+
    - SQLModel (SQLAlchemy + Pydantic)
    - JWT аутентификация
    - PostgreSQL / SQLite

  version: "2.0.0"
  contact:
    name: "API Support"
    email: "support@cinema.example.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "http://localhost:8000"
    description: "Локальный сервер разработки"
  - url: "https://api.cinema.example.com"
    description: "Продакшн сервер"
...
paths:
...
  /auth/register:
    post:
      summary: "Регистрация нового пользователя"
      description: "Создание нового аккаунта пользователя"
      tags: ["Аутентификация"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "200":
          description: "Пользователь успешно создан"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: "Ошибка валидации или пользователь уже существует"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: "Некорректные входные данные"
  /auth/login:
    post:
      summary: "Вход в систему"
      description: "Получение JWT токена для аутентификации"
      tags: ["Аутентификация"]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: "Успешная аутентификация"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: "Неверные учетные данные"
        "422":
          description: "Некорректные входные данные"

  # ============ Пользователи ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT токен аутентификации"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Ошибка валидации"
        details:
          type: object
          example:
            field: "email"
            message: "Неверный формат email"

    # ============ Пользователи ============
    UserBase:
      type: object
      required: [email, username]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        username:
          type: string
          minLength: 3
          maxLength: 100
          example: "cinemalover"
        full_name:
          type: string
          nullable: true
          example: "Иван Иванов"
    UserCreate:
      allOf:
        - $ref: "#/components/schemas/UserBase"
        - type: object
          required: [password]
          properties:
            password:
              type: string
              minLength: 8
              example: "StrongPass123"
...
    UserResponse:
      allOf:
        - $ref: "#/components/schemas/UserBase"
        - type: object
          required: [id, subscription_tier, is_active, created_at]
          properties:
            id:
              type: integer
              example: 1
            subscription_tier:
              $ref: "#/components/schemas/SubscriptionTierResponse"
            is_active:
              type: boolean
              example: true
            created_at:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00Z"
            last_login:
              type: string
              format: date-time
              nullable: true
              example: "2024-01-15T14:45:00Z"

    # ============ Аутентификация ============
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "cinemalover"
        password:
          type: string
          example: "StrongPass123"
    TokenResponse:
      type: object
      required: [access_token, token_type, user_id, username]
      properties:
        access_token:
          type: string
          example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
        token_type:
          type: string
          example: "bearer"
        user_id:
          type: integer
          example: 1
        username:
          type: string
          example: "cinemalover"
        subscription_tier:
          $ref: "#/components/schemas/SubscriptionTierResponse"

    # ============ Контент ============
...
    SubscriptionTierResponse:
      type: object
      required: [id, code, name, max_simultaneous_streams, max_video_quality]
      properties:
        id:
          type: integer
          example: 3
        code:
          type: string
          example: "premium"
        name:
          type: string
          example: "Премиум"
        price_monthly:
          type: number
          nullable: true
          example: 599
        price_yearly:
          type: number
          nullable: true
          example: 5990
        max_simultaneous_streams:
          type: integer
          example: 4
        max_video_quality:
          type: string
          example: "4K"
        description:
          type: string
          nullable: true
        features:
          type: array
          items:
            $ref: "#/components/schemas/SubscriptionFeatureResponse"
    SubscriptionFeatureResponse:
      type: object
      required: [id, feature, is_included]
      properties:
        id:
          type: integer
          example: 1
        feature:
          type: string
          example: "Без рекламы"
        is_included:
          type: boolean
          example: true
...